FastTracker/TrackEval/requirements.txt
numpy==1.18.1
scipy==1.4.1
pycocotools==2.0.2
matplotlib==3.2.1
opencv_python==4.4.0.46
scikit_image==0.16.2
pytest==6.0.1
Pillow==8.1.2
tqdm==4.64.0
tabulate
